major update mandl, UI, etc

+++ engw3dtest/shaders/basic.vert.glsl
@@ -4,7 +4,7 @@ attribute vec2 textureCoordAttribute;
 uniform mat4 mvMatrixUniform;
 uniform mat4 pMatrixUniform;
 
-varying highp vec2 vTextureCoord;
+varying mediump vec2 vTextureCoord;
 
 void main(void) {
 	gl_Position = pMatrixUniform * (mvMatrixUniform * vec4(vertexPositionAttribute, 1.0));
